move some utils to core-js
