--- packages/core-js/utils/dummy.js
@@ -1,1 +1,0 @@
-console.log('dummy util');
