+++ packages/core-js/utils/string.js
