+++ packages/deployer/subtasks/deploy-contract.js
@@ -1,4 +1,4 @@
-const logger = require('../utils/logger');
+const logger = require('@synthetixio/core-js/utils/logger');
 const {
   getContractNameFromPath,
   getContractBytecodeHash,
