+++ packages/deployer/subtasks/deploy-router.js
@@ -1,4 +1,4 @@
-const logger = require('../utils/logger');
+const logger = require('@synthetixio/core-js/utils/logger');
 const { subtask } = require('hardhat/config');
 const { TASK_COMPILE } = require('hardhat/builtin-tasks/task-names');
