+++ packages/deployer/subtasks/prepare-deployment.js
@@ -3,8 +3,8 @@ const path = require('path');
 const mkdirp = require('mkdirp');
 const { subtask } = require('hardhat/config');
 
-const prompter = require('../utils/prompter');
-const relativePath = require('../utils/relative-path');
+const prompter = require('@synthetixio/core-js/utils/prompter');
+const relativePath = require('@synthetixio/core-js/utils/relative-path');
 const autosaveObject = require('../utils/autosave-object');
 const { getDeploymentFiles } = require('../utils/deployments');
 const { SUBTASK_PREPARE_DEPLOYMENT } = require('../task-names');
