+++ packages/deployer/tasks/deploy.js
@@ -14,8 +14,8 @@ const {
   SUBTASK_VALIDATE_ROUTER,
   TASK_DEPLOY,
 } = require('../task-names');
-const logger = require('../utils/logger');
-const prompter = require('../utils/prompter');
+const logger = require('@synthetixio/core-js/utils/logger');
+const prompter = require('@synthetixio/core-js//utils/prompter');
 const { getDeploymentPaths } = require('../utils/deployments');
 const types = require('../utils/argument-types');
