+++ packages/deployer/utils/deployments.js
@@ -2,8 +2,8 @@ const fs = require('fs');
 const path = require('path');
 const glob = require('glob');
 const naturalCompare = require('string-natural-compare');
-const relativePath = require('./relative-path');
-const { capitalize, startsWithWord } = require('./string');
+const relativePath = require('@synthetixio/core-js//utils/relative-path');
+const { capitalize, startsWithWord } = require('@synthetixio/core-js//utils/string');
 
 // Regex for deployment file formats, e.g.: 2021-08-31-00-sirius.json
 const DEPLOYMENT_FILE_FORMAT = /^[0-9]{4}-[0-9]{2}-[0-9]{2}-[0-9]{2,}(?:-[a-z0-9]+)?\.json$/;
